Add project roster permissions
Applies PermissionsMixin to the Project model, so that only editors can
edit a project's roster. Additionally, changes the `get_project_role`
method in UserProfile so that it takes a project as an argument,
enabling multiple affiliations in the future. Finally, changes the
`save` method on UserProfiles so that users are automatically added to
the groups of projects that they are members of, with a default role of
researcher.