Add support for handlebars.js templating
Rendering a zotero string as a Django widget was getting very tedious to
keep up, so this adds javascript functions to generate the same type of
form client-side with handlebars.js templates

Django can't easily include raw template text, so for the moment, I get
the template scripts via an ajax request. In the future, this should be
done with some sort of {% raw_text %} template tag.