Update widget to render a zotero string
This is probably the better way to go rather than rendering everything
with javascript, since it's nice in terms of creating a consistent
experience to handle form rendering on the server side. Whichever way we
decide to move in the future, it would be helpful to keep both options
open.