Fix some errors with revisions; keep aliases around
We're going to want to reset our revision history once this insanity is
all over with. Since we started using our own user model, it's basically
impossible to run through the migrations from a clean database-- 0078
requires reversion_version to be created, but reversion_version has a
foreign key to AUTH_USER_MODEL, which is main.auth.User, which is
created in 0101. We could hack around this but I just don't know how we
could do it well.